split intro into its own file
